Use coroutine scope from project level injected service
