+++ plugins/gradle/src/org/jetbrains/plugins/gradle/service/execution/GradleDownloadProgressListener.kt
@@ -1,108 +1,41 @@
 // Copyright 2000-2023 JetBrains s.r.o. and contributors. Use of this source code is governed by the Apache 2.0 license.
 package org.jetbrains.plugins.gradle.service.execution
 
-import com.intellij.build.events.ProgressBuildEvent
+import com.intellij.openapi.components.Service
+import com.intellij.openapi.components.service
 import com.intellij.openapi.externalSystem.model.task.ExternalSystemTaskId
 import com.intellij.openapi.externalSystem.model.task.ExternalSystemTaskNotificationEvent
-import com.intellij.openapi.externalSystem.model.task.ExternalSystemTaskNotificationListenerAdapter
-import com.intellij.openapi.externalSystem.model.task.event.ExternalSystemBuildEvent
-import com.intellij.openapi.externalSystem.service.internal.ExternalSystemTaskProgressIndicatorUpdater.getText
-import com.intellij.openapi.externalSystem.service.notification.ExternalSystemProgressNotificationManager
-import com.intellij.openapi.externalSystem.util.ExternalSystemBundle
-import com.intellij.openapi.progress.*
-import com.intellij.openapi.project.Project
 import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.channels.Channel
-import kotlinx.coroutines.launch
-import org.jetbrains.annotations.Nls
-import org.jetbrains.plugins.gradle.util.GradleConstants
 
 /**
  * Listens to Gradle download progress events and shows a download progress as progress indicator for every artifact.
  */
-class GradleDownloadProgressListener(taskId: ExternalSystemTaskId): ExternalSystemTaskNotificationListenerAdapter() {
+interface GradleDownloadProgressListener {
 
-  private val project: Project? = taskId.findProject()
-  private val myProgressIndicators: MutableMap<Any, DownloadEventProgressIndicator> = mutableMapOf()
-  private val cs = CoroutineScope(Dispatchers.Main)
-
-  init {
-    ExternalSystemProgressNotificationManager.getInstance().addNotificationListener(taskId, this)
-  }
-
-  fun updateProgressIndicator(event: ExternalSystemTaskNotificationEvent) {
-    if (project != null && event is ExternalSystemBuildEvent && event.buildEvent is ProgressBuildEvent) {
-      val buildEvent = event.buildEvent as ProgressBuildEvent
-      val progressIndicator = myProgressIndicators.computeIfAbsent(event.buildEvent.id) {
-        DownloadEventProgressIndicator(project, cs, event.description)
-      }
-      progressIndicator.updateProgressIndicator(buildEvent)
-    }
-  }
-
-  fun stopProgressIndicator(event: ExternalSystemTaskNotificationEvent) {
-    if (event is ExternalSystemBuildEvent && event.buildEvent is ProgressBuildEvent) {
-      myProgressIndicators.remove(event.buildEvent.id)?.stop()
-    }
-  }
-
-  /**
-   * We listen to `onEnd` so we cleanup everything even if Gradle crashes.
-   */
-  override fun onEnd(id: ExternalSystemTaskId) {
-    // Stop all indicators that might be still visible, e.g. on Daemon crash
-    myProgressIndicators.forEach { (_, indicator) -> indicator.stop() }
-    myProgressIndicators.clear()
-    ExternalSystemProgressNotificationManager.getInstance().removeNotificationListener(this)
-  }
-
-  private class DownloadEventProgressIndicator(private val project: Project, private val cs: CoroutineScope, title: String) {
-
-    private val channel: Channel<ProgressBuildEvent> = Channel()
-
-    init {
-      startProgressIndicator(title)
-    }
-
-    fun updateProgressIndicator(buildEvent: ProgressBuildEvent) {
-      cs.launch {
-        if (!channel.isClosedForSend) {
-          channel.send(buildEvent)
+  companion object {
+    @JvmStatic
+    fun newListener(taskId: ExternalSystemTaskId): GradleDownloadProgressListener =
+      when (val project = taskId.findProject()) {
+        null -> NoOpGradleDownloadProgressListener()
+        else -> {
+          val scopeProvider = project.service<GradleProgressCoroutineScopeProvider>()
+          GradleDownloadProgressListenerImpl(taskId, project, scopeProvider.cs)
         }
       }
-    }
+  }
 
-    private fun startProgressIndicator(title: String) {
-      cs.launch {
-        withBackgroundProgress(project, wrapText(title), TaskCancellation.nonCancellable()) {
-          var previousFraction = 0.0
-          var previousStep: ProgressReporter? = null
-          var statusEvent = channel.receiveCatching().getOrNull()
-          while (statusEvent != null && statusEvent.fraction < 1.0) {
-            val fraction = statusEvent.fraction
-            if (fraction > previousFraction) {
-              // End previous step only when we start a new one so that we see bytes progress in details.
-              // That also means that in reality indicator will show fraction progress for 1 step behind, but that should not be noticeable.
-              previousStep?.close()
-              val text = getText(title, statusEvent.progress, statusEvent.total, statusEvent.unit) { wrapText(it) }
-              previousStep = progressReporter?.step(fraction, text)
-              previousFraction = fraction
-            }
-            statusEvent = channel.receiveCatching().getOrNull()
-          }
-        }
-      }
-    }
+  fun updateProgressIndicator(event: ExternalSystemTaskNotificationEvent)
+  fun stopProgressIndicator(event: ExternalSystemTaskNotificationEvent)
+}
 
-    private fun wrapText(text: String): @Nls String =
-      ExternalSystemBundle.message("progress.update.text", GradleConstants.SYSTEM_ID.readableName, text)
+@Service(Service.Level.PROJECT)
+class GradleProgressCoroutineScopeProvider(val cs: CoroutineScope)
 
-    fun stop() {
-      channel.close()
-    }
+internal
+class NoOpGradleDownloadProgressListener : GradleDownloadProgressListener {
+  override fun updateProgressIndicator(event: ExternalSystemTaskNotificationEvent) {
+  }
 
-    private val ProgressBuildEvent.fraction
-      get(): Double = progress.toDouble() / total
+  override fun stopProgressIndicator(event: ExternalSystemTaskNotificationEvent) {
   }
 }
